flask-login-app/requirements.txt
Flask-Bcrypt
Flask
Flask-Bootstrap
Flask-SQLAlchemy
Flask-Login
WTForms
SQLAlchemy
Flask-WTF
email_validator
bs4
requests
lxml